contracts/projects/VoiceVault/pyproject.toml
[tool.poetry]
name = "VoiceVault"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["ghoshvidip26 <ghoshvidip26@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
beaker-pyteal = "^1.1.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
